sredeno se odven welcome

+++ soccerstats/app/Http/Controllers/GameController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\GameRequest;
 use App\Models\Game;
 use App\Models\Game_player;
 use App\Models\Team;
@@ -10,75 +11,42 @@ use Illuminate\Support\Facades\Auth;
 
 class GameController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
     public function index()
     {
-        return view('match.index', ['matches' => Game::all()]);
+        return view('match.index', ['matches' => Game::orderBy('schaduled_at')->paginate(10)]);
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
     public function create()
     {
         return view('match.create', ['teams' => Team::all()]);
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
-    public function store(Request $request)
+    public function store(GameRequest $request)
     {
-        $request->validate([
-            
-            'home_team' => 'required|different:guest_team',
-            'guest_team' => 'required|different:home_team',
-            'schaduled_at' => 'required|date',
-        ]);
-
         $match = Game::make();
         $match->user_id = Auth::user()->id;
         $match->home_team = $request->home_team;
         $match->guest_team = $request->guest_team;
         $match->schaduled_at = $request->schaduled_at;
         $match->save();
-
         foreach(Team::find($match->home_team)->players as $player) {
             Game_player::create([
                 'game_id' => $match->id,
                 'player_id' => $player->id,
             ]);
         }
-
         foreach(Team::find($match->guest_team)->players as $player) {
             Game_player::create([
                 'game_id' => $match->id,
                 'player_id' => $player->id,
             ]);
         }
-
         return redirect()->back()->with('status', 'Game created successfully');
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  \App\Models\Match  $match
-     * @return \Illuminate\Http\Response
-     */
     public function show($id)
     {
         $match = Game::find($id);
-
         $homePlayers = [
             'played' => [],
             'bench' => [],
@@ -87,7 +55,6 @@ class GameController extends Controller
             'played' => [],
             'bench' => [],
         ];
-        
         foreach($match->homeTeam->players as $player) {
             if(Game_player::where('game_id', $match->id)->where('player_id', $player->id)->first()->has_played) {
                 $homePlayers['played'][] = $player;
@@ -95,7 +62,6 @@ class GameController extends Controller
                 $homePlayers['bench'][] = $player;
             }
         }
-
         foreach($match->guestTeam->players as $player) {
             if(Game_player::where('game_id', $match->id)->where('player_id', $player->id)->first()->has_played) {
                 $guestPlayers['played'][] = $player;
@@ -103,63 +69,32 @@ class GameController extends Controller
                 $guestPlayers['bench'][] = $player;
             }
         }
-
         return view('match.show', ['match' => $match, 'homePlayers' => $homePlayers, 'guestPlayers' => $guestPlayers]);
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  \App\Models\Match  $match
-     * @return \Illuminate\Http\Response
-     */
     public function edit($id)
     {
         return view('match.edit', ['match' => Game::find($id), 'teams' => Team::all(), 'home_players' => Team::find(Game::find($id)->home_team)->players, 'guest_players' => Team::find(Game::find($id)->guest_team)->players]);
     }
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Models\Match  $match
-     * @return \Illuminate\Http\Response
-     */
-    public function update(Request $request, $id)
+    public function update(GameRequest $request, $id)
     {
-        $request->validate([
-            'home_team' => 'required|different:guest_team',
-            'guest_team' => 'required|different:home_team',
-            'schaduled_at' => 'required|date',
-        ]);
-
-        $change_in_teams = false;
-
         $match = Game::find($id);
-
         if($match->home_team != $request->home_team || $match->guest_team != $request->guest_team) {
             Game_player::where('game_id', $match->id)->delete();
-
             foreach(Team::find($request->home_team)->players as $player) {
                 Game_player::create([
                     'game_id' => $match->id,
                     'player_id' => $player->id,
                 ]);
             }
-    
             foreach(Team::find($request->guest_team)->players as $player) {
                 Game_player::create([
                     'game_id' => $match->id,
                     'player_id' => $player->id,
                 ]);
             }
-
-            $change_in_teams = true;
         }
-
-
-        
-
         $match->user_id = Auth::user()->id;
         $match->home_team = $request->home_team;
         $match->guest_team = $request->guest_team;
@@ -168,14 +103,11 @@ class GameController extends Controller
         $match->guest_score = $request->guest_score;
         $match->is_finished = $request->is_finished == 'on' ? true : false;
         $match->update();
-
-        
         foreach($match->players as $player) {
             Game_player::where('game_id', $match->id)->where('player_id', $player->id)->update([
                 'has_played' => false,
             ]);
         }
-
         if($request->home_players) {
             foreach($request->home_players as $player_id => $has_played) {
                 Game_player::where('game_id', $match->id)->where('player_id', $player_id)->update([
@@ -183,7 +115,6 @@ class GameController extends Controller
                 ]);
             }
         }
-
         if($request->guest_players) {
             foreach($request->guest_players as $player_id => $has_played) {
                 Game_player::where('game_id', $match->id)->where('player_id', $player_id)->update([
@@ -191,21 +122,18 @@ class GameController extends Controller
                 ]);
             }
         }
-
         return redirect()->back()->with('status', 'Game updated successfully');
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  \App\Models\Match  $match
-     * @return \Illuminate\Http\Response
-     */
-    public function destroy($id)
+    public function destroy(Request $request, $id)
     {
+        // dd($request->all());
 
-        Game::find($id)->delete();
-
-        return redirect()->back()->with('status', 'Match deleted successfully');
+        if($request->has('confirm_deletion')) {
+            Game::find($id)->delete();
+            return redirect()->back()->with('status', 'Match deleted successfully');
+        } else {
+            return redirect()->back()->with('error', 'Please confirm deletion of the match');
+        }
     }
 }
