+++ soccerstats/app/Http/Controllers/TeamController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\TeamRequest;
 use App\Models\Player_team;
 use App\Models\Team;
 use Illuminate\Http\Request;
@@ -9,104 +10,54 @@ use Illuminate\Support\Facades\Auth;
 
 class TeamController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
     public function index()
     {
-        return view('team.index', ['teams' => Team::all()]);
+        return view('team.index', ['teams' => Team::paginate(10)]);
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
     public function create()
     {
         return view('team.create');
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
-    public function store(Request $request)
+    public function store(TeamRequest $request)
     {
-        $request->validate([
-            'name' => 'required|string|max:255',
-            'year_founded' => 'required',
-        ]);
-
         $team = Team::make();
         $team->user_id = Auth::user()->id;
         $team->name = $request->name;
         $team->year_founded = $request->year_founded;
         $team->save();
-
         return redirect()->back()->with('status', 'Team successfully added.');
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  \App\Models\Team  $team
-     * @return \Illuminate\Http\Response
-     */
     public function show($id)
     {
         return view('team.show', ['team' => Team::find($id)]);
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  \App\Models\Team  $team
-     * @return \Illuminate\Http\Response
-     */
     public function edit($id)
     {
         return view('team.edit', ['team' => Team::find($id)]);
     }
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Models\Team  $team
-     * @return \Illuminate\Http\Response
-     */
-    public function update(Request $request, $id)
+    public function update(TeamRequest $request, $id)
     {
-        $request->validate([
-            'name' => 'required|string|max:255',
-            'year_founded' => 'required',
-        ]);
-
         $team = Team::find($id);
         $team->user_id = Auth::user()->id;
         $team->name = $request->name;
         $team->year_founded = $request->year_founded;
         $team->save();
-
         return redirect()->back()->with('status', 'Team successfully added.');
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  \App\Models\Team  $team
-     * @return \Illuminate\Http\Response
-     */
-    public function destroy($id)
+    public function destroy(Request $request, $id)
     {
-        Player_team::where('team_id', $id)->delete();
-        Team::find($id)->delete();
-
-        return redirect()->back()->with('status', 'Team was deleted');
+        if($request->has('confirm_deletion')) {
+            Player_team::where('team_id', $id)->delete();
+            Team::find($id)->delete();
+            return redirect()->back()->with('status', 'Team was deleted');
+        } else {
+            return redirect()->back()->with('error', 'Please confirm deletion of the team');
+        }
     }
 }
